requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.1
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
Pillow==7.0.0
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
